// src/themes/types/theme.ts
import {
  type PaletteColorOptions,
  type SimplePaletteColorOptions,
} from "@mui/material/styles";

export interface PaletteThemeProps {
  primary: SimplePaletteColorOptions;
  secondary: SimplePaletteColorOptions;
  error: SimplePaletteColorOptions;
  warning: SimplePaletteColorOptions;
  info: SimplePaletteColorOptions;
  success: SimplePaletteColorOptions;
  grey: PaletteColorOptions;
}

export interface CustomShadowProps {
  button: string;
  text: string;
  z1: string;
  primary: string;
  primaryButton: string;
  secondary: string;
  secondaryButton: string;
  error: string;
  errorButton: string;
  warning: string;
  warningButton: string;
  info: string;
  infoButton: string;
  success: string;
  successButton: string;
  grey: string;
  greyButton: string;
}

export enum ThemeMode {
  LIGHT = "light",
  DARK = "dark",
}